% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreTargets.R
\name{exploreTargets}
\alias{exploreTargets}
\title{Explore targets in real data}
\usage{
exploreTargets(object, n.target = 2, n.ions = 3, rt.lim = NULL)
}
\arguments{
\item{iTERPSet}{S4 structure class ITERP}

\item{mz.num}{number of ions we want to determine for each targe}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
exploreTargets(object=iTERPSet, n.target = 2, n.ions=3, rt.lim=c(2.5,2.8))

}
