% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{getResults}
\usage{
getResults(object, type = c("A", "I"))
}
\arguments{
\item{n.ions}{number of ions we want to quantify for each targe}
}
\description{
This function retrieves Area, Intensity and Extracted Ions Chromatograms for each Target
in real data. They are stores in slot R.
}
\examples{
results <- getResults(object, type=c("I"))
results <- getResults(object, type=c("A"))
}
