% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateTargets.R
\name{integrateTargets}
\alias{integrateTargets}
\title{Integrates targets in real data}
\usage{
integrateTargets(object, n.ions = 3)
}
\arguments{
\item{n.ions}{number of ions we want to quantify for each targe}
}
\description{
This function integrates the n.ions for each target and calculates Area, Intensity and Extracted Ions Chromatograms for each Target
in real data. They are stores in slot R.
}
\examples{
iTERPSet <- integrateTargets(object=iTERPSet, n.ions = 3)
}
