% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateTICs.R
\name{integrateTIC}
\alias{integrateTIC}
\title{Integrates TICs}
\usage{
integrateTIC(object)
}
\arguments{
\item{iTERPSet}{S4 structure class ITERP}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
ATIC <- integrateTIC(object=iTERPSet)

}
