% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIST2Targets2_2.R
\docType{class}
\name{iTERP-class}
\alias{iTERP-class}
\alias{iTERPSet}
\title{NIST2Targets}
\usage{
iTERPSet <- NIST2targets()
}
\description{
Reads .MSP files for a directory and saves in to a target matrix
}
\details{
aaaaaaaaaaaaaaaaaaaaaaaaaaaa
}
