% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateTargets2.R
\name{locateTargets}
\alias{locateTargets}
\title{Locate Targets in real data}
\usage{
locateTargets(object, scan.window = 10)
}
\arguments{
\item{scan.window}{is the rt window in scans for maximum width of target peaks}

\item{iTERPSet}{S4 structure class ITERP}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
iTERPSet <- locateTargets(iTERPSet,scan.window = 10)

}
