% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEICs.R
\name{plotEICs}
\alias{plotEICs}
\title{plotEICs}
\usage{
plotEICs(object, n.target = 1)
}
\arguments{
\item{iTERPSet}{S4 structure class ITERP}

\item{n.targ}{is the target columns in iTERPSet@Targets.spectra}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
plotEICs(object=iTERPSet, n.target = 1)
}
