% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMSLibraryTargets.R
\name{plotMSLibraryTargets}
\alias{plotMSLibraryTargets}
\title{Locate Targets in real data}
\usage{
plotMSLibraryTargets(object, n.target = 2, mz.min = 35, mz.max = 300)
}
\arguments{
\item{iTERPSet}{S4 structure class ITERP}

\item{scan.window}{is the rt window in scans for maximum width of target peaks}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
iTERPSet <- locateTargets(iTERPSet,scan.window = 10)

}
