% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTICs.R
\name{plotTICs}
\alias{plotTICs}
\title{Locate Targets in real data}
\usage{
plotTICs(object, sample2plot = NULL, rt.lim = NULL,
  sample.class = NULL)
}
\arguments{
\item{sample2plot}{is  the number of dample to plot. If it is null an overlay of all TICS is iTERPset
is going to be displayed}

\item{rt.lim}{is the rt.limit you want to plot}

\item{iTERPSet}{S4 structure class ITERP}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
plotTICs(object=iTERPSet, sample2plot=6,rt.lim=c(2.5, 2.8))
plotTICs(object=iTERPSet,rt.lim=c(2.5, 2.7), sample.class=NULL)
}
