% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runAll}
\alias{runAll}
\title{Locate Targets in real data}
\usage{
runAll(object, rt.window = 3, n.ions = 3)
}
\arguments{
\item{rt.window}{is the rt window in seconds for maximum width of target peaks}

\item{n.ions}{are the number of top-intense ions that you want to integrate}

\item{iTERPSet}{S4 structure class ITERP}
}
\description{
This function is aimed to establish the rt of target terpenoids in real data
}
\examples{
iTERPSet <- runAll(iTERPSet,rt.window = 3, n.ions=3)

}
